requirements.txt
SQLAlchemy==1.4.25
PyMySQL==1.0.2
redis==3.5.3
pandas==1.3.3
celery==5.1.2
python-dotenv==0.19.1